[FIX] 몇몇 코드 문제해결중

+++ src/main/java/com/ohgiraffers/projectgin/model/repository/MemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 @Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
     Optional<MemberEntity> findMemberEntityByMemberId(String memberId);
     boolean existsByMemberId(String userId);
